[Bucket] reset as double rather than int (derp)

+++ Oxide.Ext.Discord/REST/Bucket.cs
@@ -16,7 +16,7 @@
 
         public int Remaining { get; set; }
 
-        public int Reset { get; set; }
+        public double Reset { get; set; }
 
         public bool Initialized { get; private set; } = false;
